+++ NEWS.md
@@ -3,6 +3,7 @@
 - Changed the default `min_time` (minimum seconds elapsed per query) value to 1 (60 queries per minute) for the Location IQ service ([#166](https://github.com/jessecambon/tidygeocoder/issues/166)).
 - Updated default Geocodio API URL from version 1.6 to 1.7.
 - Fixed code and documentation that incorrectly referred to `mapquest_open` as `mapbox_open`.
+- An error is now thrown if an `api_options` parameter is not compatible with the specified `method`.
 - A message is now displayed warning that `flatten=FALSE` is ignored for Geocodio and Mapquest (the output of these services requires flattening to avoid errors).
 
 # tidygeocoder 1.0.5